[WIP] Move `fused_chunk` into general simple gla
